+++ pyffi/object_models/mex/__init__.py
@@ -5,7 +5,7 @@ into Python classes.
 
 # ***** BEGIN LICENSE BLOCK *****
 #
-# Copyright (c) 2007-2009, Python File Format Interface
+# Copyright (c) 2007-2010, Python File Format Interface
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without